JB23AT2-32-Get-Hotel-room - finished

+++ src/main/java/com/aacademy/moonlight/dto/hotel/RoomRequest.java
@@ -20,7 +20,7 @@ public class RoomRequest {
 
     @NotNull
     @Min(value = 1)
-    private Long roomNumber;
+    private Integer roomNumber;
 
     @NotNull
     @Min(value = 100)
